split the env schemas
