+++ pages/api/auth/[...nextauth].js
@@ -1,7 +1,7 @@
 import NextAuth from "next-auth";
 import GithubProvider from "next-auth/providers/github";
 
-import { env } from "@config/envConfig";
+import { serverEnv } from "@config/schemas/serverSchema";
 import DbAdapter from "./db-adapter";
 import connectMongo from "@config/mongo";
 
@@ -9,8 +9,8 @@ export const authOptions = {
   adapter: DbAdapter(connectMongo),
   providers: [
     GithubProvider({
-      clientId: env.GITHUB_ID,
-      clientSecret: env.GITHUB_SECRET,
+      clientId: serverEnv.GITHUB_ID,
+      clientSecret: serverEnv.GITHUB_SECRET,
       profile(profile) {
         return {
           id: profile.id.toString(),
