+++ pages/playground.js
@@ -6,10 +6,10 @@ import Modal from "@components/Modal";
 import Input from "@components/form/Input";
 import UserPage from "@components/user/UserPage";
 import Notification from "@components/Notification";
-import { env } from "@config/envConfig";
+import { clientEnv } from "@config/schemas/clientSchema";
 
 export default function Playground() {
-  const BASE_URL = env.NEXT_PUBLIC_BASE_URL;
+  const BASE_URL = clientEnv.NEXT_PUBLIC_BASE_URL;
   const defaultJson = `{
     "name": "Your Name",
     "bio": "Write a short bio about yourself",
