+++ tests/setup/auth.js
@@ -1,5 +1,5 @@
 import connectMongo from "@config/mongo";
-import { env } from "@config/envConfig";
+import { serverEnv } from "@config/schemas/serverSchema";
 import { encode } from "next-auth/jwt";
 
 import { User, Session, Account } from "@models/index";
@@ -17,7 +17,7 @@ const login = async (browser) => {
       username: "eddiejaoude",
       id: "22222222",
     },
-    secret: env.NEXTAUTH_SECRET,
+    secret: serverEnv.NEXTAUTH_SECRET,
   });
 
   let testUser;
